Update firststeps.mdx
Is registration also necessary for the test version?
I don't know what this means: "A rework of the API to meet the requirement of SciFinderN will be provided as soon as possible." Other users can not use the SciFinder modul due to university restrictions oder is the ELN having restrictions?
I replaced "Instance admin" by "Server Admin", I am not shure if this is the correct description, but it would be helpful for me to have one term for the admin resposible for "my" server. Maybe you can go through the pages starting from "Browser and access" to unify the term used for that admin.